requirements.txt
numpy
pandas
matplotlib
seaborn
tqdm
torch
torchutils
tokenizers
torchmetrics
coremltools
python-Levenshtein
anytree
h5py
torchtext
sqlalchemy

ipython
